Move DateRange to own class
