+++ src/main/angular/app/date_select/date_select.component.html
@@ -1,6 +1,6 @@
 <mat-form-field appearance="fill">
   <mat-label>Enter a date range</mat-label>
-  <mat-date-range-input [min]="possibleRange.start" [max]="possibleRange.end"
+  <mat-date-range-input [min]="possibleRange.getStart()" [max]="possibleRange.getEnd()"
     [rangePicker]="picker">
     <input matStartDate placeholder="Start date" [value]="currentMin"
       (dateChange)="changeStart($event)">
